add cells plot, move things to components folder
