+++ src/packages/@ncigdc/modern_components/SCRNASeq/plots/index.js
@@ -1,3 +1,4 @@
 export { default as MedianGenesPlot } from './MedianGenesPlot';
 export { default as SCRNASeqPlot } from './SCRNASeqPlot';
 export { default as SequencingSaturationPlot } from './SequencingSaturationPlot';
+export { default as SummaryCellsPlot } from './SummaryCellsPlot';
