+++ src/packages/@ncigdc/modern_components/SCRNASeq/tables/SummaryTable.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import EntityPageHorizontalTable from '@ncigdc/components/EntityPageHorizontalTable';
+import { SummaryCellsPlot } from '../plots';
 
 const SummaryTable = ({ containerStyle, header, rows }) => (
   <div key={header} style={containerStyle}>
@@ -12,6 +13,7 @@ const SummaryTable = ({ containerStyle, header, rows }) => (
       >
       {header}
     </h3>
+    {header === 'Cells' && <SummaryCellsPlot />}
     <EntityPageHorizontalTable
       data={rows.map(([key, value]) => ({
         key,
